+++ internal/repository/session/repository.go
@@ -3,12 +3,12 @@ package session
 type Token string
 type Session struct {
 	Email     string
-	ExpiresAt int64
+	ExpiresAt uint
 }
 
 type Repository interface {
 	NewSession(email string) (string, error)
 	GetSession(token Token) (string, error)
 	DeleteSession(token Token) error
-	Expiring() int64
+	Expiring() uint
 }
